requirements.txt
openai==0.28.1
flask==3.0.0
python-dotenv==1.0.0
email-validator==2.1.0
requests==2.31.0
werkzeug==3.0.1
pytest==7.4.3
pytest-cov==4.1.0